refactor(sync): move the box-wins untracked-overlay classifier into a concern
Phase 6 (1/2). Move classifyUntrackedOverlay + NON_REGULAR_TOKEN — the pure
box-wins content-hash policy the reconciler contract names as its first
consumer — into a new sandbox-core/sync/concerns/git.ts. Docker in-box-git.ts
re-exports it (shim — the docker test + the resync caller untouched). New
git-concern.test.ts (4) mirrors the docker classify cases (double-parity).

Gate: build + typecheck + lint green; core (85, incl. git-concern) + docker
(297, incl. classify via the re-export) green.
