+++ packages/sandbox-core/src/sync/index.ts
@@ -58,6 +58,7 @@ export {
   type CredentialAgentKind,
   type ExtractCredentialsOptions,
 } from './concerns/credentials.js';
+export { classifyUntrackedOverlay, NON_REGULAR_TOKEN } from './concerns/git.js';
 export {
   BOX_WORKFLOWS_DIR,
   BOX_MEMORY_DIR,
